+++ src/main/java/com/idigitronics/IDigi/request/vo/UserManagementRequestVO.java
@@ -11,7 +11,7 @@ public class UserManagementRequestVO {
 	
 	private int communityID;
 	private int blockID;
-	private int customerID;
+	private long customerID;
 	private String userID;
 	private String userName;
 	private String userPassword;
@@ -90,10 +90,10 @@ public class UserManagementRequestVO {
 	public void setOldPassword(String oldPassword) {
 		this.oldPassword = oldPassword;
 	}
-	public int getCustomerID() {
+	public long getCustomerID() {
 		return customerID;
 	}
-	public void setCustomerID(int customerID) {
+	public void setCustomerID(long customerID) {
 		this.customerID = customerID;
 	}
 	public String getCustomerUniqueID() {
